refactor reservation
