+++ app.js
@@ -9,7 +9,7 @@ const { graphqlUploadExpress } = require('graphql-upload');
 const expressPlayground = require('graphql-playground-middleware-express').default;
 
 const schema = require('./modules');
-const router = require('./http/router');
+const { path, router } = require('./http');
 
 /**
  *
@@ -22,7 +22,7 @@ module.exports = (container) => {
   app.use(bodyParser.json());
   app.use(scopePerRequest(container));
   app.use(cors());
-  app.use('/api', router);
+  app.use(path.HTTP_API_ROOT, router);
   app.use('/graphql', graphqlUploadExpress(config.graphqlUploadExpress), graphqlHTTP((req) => ({
     schema,
     graphiql: config.app.env === 'development',
