+++ http/index.js
@@ -1,0 +1,4 @@
+const router = require('./router');
+const path = require('./path');
+
+module.exports = { router, path };
