+++ http/path.js
@@ -1,0 +1,3 @@
+const HTTP_API_ROOT = '/api';
+
+module.exports = { HTTP_API_ROOT };
