+++ http/reservation/path.js
@@ -1,0 +1,3 @@
+const ZALO_RESERVATION_CONFIRMATION = '/zalo/reservation/confirmation';
+
+module.exports = { ZALO_RESERVATION_CONFIRMATION };
